add delete permission

+++ coursera/coursera_api/courses/views.py
@@ -238,6 +238,7 @@ class CourseDetail(APIView):
 
     def delete(self, request, pk, format=None):
         course = self.get_object(pk)
+        self.check_object_permissions(self.request, course)
         course.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
